refactor: Move props from .tsx file to type.ts
