+++ src/components/atoms/BgText/type.ts
@@ -1,1 +1,5 @@
 export type Size = 'lg' | 'sm'
+
+export interface Props {
+  size: Size
+}
