+++ src/components/atoms/BlogCategory/type.ts
@@ -1,1 +1,7 @@
 export type Size = 'lg' | 'sm'
+
+export interface Props {
+  size?: Size
+  icon: string
+  label: string
+}
