+++ src/components/atoms/SectionTitle/type.ts
@@ -1,1 +1,6 @@
+import { TextProps } from '@chakra-ui/react'
+
 export type Size = 'lg' | 'sm'
+export interface Props extends TextProps {
+  size: Size
+}
